+++ source/Public/Get-HostTag.ps1
@@ -3,21 +3,21 @@ function Get-HostTag {
     .SYNOPSIS
         Creates a host name or tag based on predetermined criteria for as many as 999 hosts at a time.
     .DESCRIPTION
-        A longer description of the function, its purpose, common use cases, etc.
+        This function generates custom host names based on predetermined criteria. The criteria includes physical/virtual nature of the host, a 2-3 letter prefix, system OS, and device function. The name is composed of the prefix, a code for physical/virtual nature of the host, a code for the system OS, a code for device function, and a three-digit number. The name is no longer than 15 characters. The function can generate up to 999 names at a time.
     .EXAMPLE
-        Get-HostTag -PhysicalOrVirtual Physical -Prefix "CSN" -SystemOS 'Windows Server' -DeviceFunction 'Application Server' -HostCount 5
+        PS C:\> Get-HostTag -PhysicalOrVirtual Physical -Prefix "CSN" -SystemOS 'Windows Server' -DeviceFunction 'Application Server' -HostCount 5
+        Returns:
             CSN-PWSVAPP001
             CSN-PWSVAPP002
             CSN-PWSVAPP003
             CSN-PWSVAPP004
             CSN-PWSVAPP005
-        This creates the name of the host under 15 characters and numbers them. Prefix can be 2-3 characters.
     .PARAMETER PhysicalOrVirtual
-        Tab through selections to add 'P' or 'V' for physical or virtual to host tag.
+        The physical or virtual nature of the host. Enter "P" for physical and "V" for virtual.
     .PARAMETER Prefix
-        Enter the 2-3 letter prefix. Good for prefixing company initials, locations, or other.
+        The 2-3 letter prefix used in the host name.
     .PARAMETER SystemOS
-        Use tab to cycle through the following options:
+        The operating system of the host. Use tab to cycle through the following options:
             "Cisco ASA", "Android", "Apple IOS",
             "Dell Storage Center", "MACOSX",
             "Dell Power Edge", "Embedded", "Embedded Firmware",
@@ -27,7 +27,7 @@ function Get-HostTag {
             "ProxMox", "Windows Workstation", "Windows Server",
             "Windows Server Core", "Generic OS", "Generic HyperVisor"
     .PARAMETER DeviceFunction
-        Use tab to cycle through the following options:
+        The function of the device. Use tab to cycle through the following options:
             "Application Server", "Backup Server", "Directory Server",
             "Email Server", "Firewall", "FTP Server",
             "Hypervisor", "File Server", "NAS File Server",
@@ -36,32 +36,40 @@ function Get-HostTag {
             "Management", "Blade Enclosure", "Blade Enclosure Switch",
             "SAN specific switch", "General server/Network switch", "Generic Function Device"
     .PARAMETER HostCount
-        Enter a number from 1 to 999 for how many hostnames you'd like to create.
+        The number of host names to generate. Must be between 1 and 999.
     #>
+    # Define the output type of the function
     [OutputType([string[]])]
+    # Define the binding for the cmdlet
     [CmdletBinding()]
+    # Define the parameters for the function
     param (
+        # Define the first parameter, which is mandatory
         [Parameter(
-            MandaTory = $true,
-            Position = 0,
-            HelpMessage = 'Enter 2 character site code or prefix for your devices',
-            ValueFromPipelineByPropertyName = $true
+            MandaTory = $true,   # This parameter is mandatory
+            Position = 0,       # This parameter should be the first one in the list
+            HelpMessage = 'Enter 2 character site code or prefix for your devices',  # Help message for the parameter
+            ValueFromPipelineByPropertyName = $true  # This parameter can be piped to
         )]
-        [ValidateSet("Physical", "Virtual")]
-        [string]$PhysicalOrVirtual,
+        [ValidateSet("Physical", "Virtual")]  # This parameter can only have these values
+        [string]$PhysicalOrVirtual,          # The variable that will hold the value of this parameter
+
+        # Define the second parameter, which is mandatory
         [Parameter(
-            MandaTory = $true,
-            Position = 1,
-            HelpMessage = 'Enter 2 to 3 character site code or prefix for your devices',
-            ValueFromPipelineByPropertyName = $true
+            MandaTory = $true,   # This parameter is mandatory
+            Position = 1,       # This parameter should be the second one in the list
+            HelpMessage = 'Enter 2 to 3 character site code or prefix for your devices',  # Help message for the parameter
+            ValueFromPipelineByPropertyName = $true  # This parameter can be piped to
         )]
-        [ValidateLength(2, 3)]
-        [string]$Prefix,
+        [ValidateLength(2, 3)]  # This parameter can only have a value of length 2 or 3
+        [string]$Prefix,        # The variable that will hold the value of this parameter
+
+        # Define the third parameter, which is mandatory
         [Parameter(
-            MandaTory = $true,
-            Position = 2,
-            HelpMessage = 'Tab complete to pick from a list of System OSs',
-            ValueFromPipelineByPropertyName = $true
+            MandaTory = $true,   # This parameter is mandatory
+            Position = 2,       # This parameter should be the third one in the list
+            HelpMessage = 'Tab complete to pick from a list of System OSs',  # Help message for the parameter
+            ValueFromPipelineByPropertyName = $true  # This parameter can be piped to
         )]
         [ValidateSet(
             "Cisco ASA", "Android", "Apple IOS",
@@ -72,33 +80,21 @@ function Get-HostTag {
             "Nutanix", "TrueNas", "FreeNas",
             "ProxMox", "Windows Workstation", "Windows Server",
             "Windows Server Core", "Generic OS", "Generic HyperVisor"
-        )]
-        [string]$SystemOS,
+        )]  # This parameter can only have values from this list
+        [string]$SystemOS,     # The variable that will hold the value of this parameter
+
+        # Define the fourth parameter, which is mandatory
         [Parameter(
-            MandaTory = $true,
-            Position = 3,
-            HelpMessage = 'Tab complete to pick from a list of Device Functions',
-            ValueFromPipelineByPropertyName = $true
+            Position = 3,       # This parameter should be the fourth one in the list
+            HelpMessage = 'Enter the number of host names you want to create between 1 and 254',  # Help message for the parameter
+            ValueFromPipelineByPropertyName = $true  # This parameter can be piped to
         )]
-        [ValidateSet(
-            "Application Server", "Backup Server", "Directory Server",
-            "Email Server", "Firewall", "FTP Server",
-            "Hypervisor", "File Server", "NAS File Server",
-            "Power Distribution Unit", "Redundant Power Supply", "SAN Appliance",
-            "SQL Server", "Uninteruptable Power Supply", "Web Server",
-            "Management", "Blade Enclosure", "Blade Enclosure Switch",
-            "SAN specific switch", "General server/Network switch", "Generic Function Device"
-        )]
-        [string]$DeviceFunction,
-        [Parameter(
-            Position = 4,
-            HelpMessage = 'Enter the number of host names you want to create between 1 and 254',
-            ValueFromPipelineByPropertyName = $true
-        )]
-        [ValidateRange(1, 999)]
-        [int]$HostCount = 1
+        [ValidateRange(1, 999)]  # This parameter can only have a value between 1 and 999
+        [int]$HostCount = 1     # The variable that will hold the value of this parameter
     )
+    # Define the begin block
     begin {
+        # The DeviceFunction parameter
         switch ($DeviceFunction) {
             "Application Server" { $DFunction = "APP" }
             "Backup Server" { $DFunction = "BAK" }
@@ -122,6 +118,7 @@ function Get-HostTag {
             "General Server/Network Switch" { $DFunction = "SW-SVR" }
             Default { $DFunction = "XDV" }
         }
+        # Set a value based on the value of the SystemOS parameter
         switch ($SystemOS) {
             "Cisco ASA" { $OSTxt = "ASA" }
             "Android" { $OSTxt = "DRD" }
@@ -147,20 +144,25 @@ function Get-HostTag {
             "Generic OS" { $OSTxt = "GOS" }
             Default { $DFunction = "GHV" }
         }
+        # Set a value based on the value of the PhysicalOrVirtual parameter
         switch ($PhysicalOrVirtual) {
             "Physical" { $DevType = "P" }
             Default { $DevType = "V" }
         }
     }
+    # Define the process block
     process {
         $OutPut = @()
         1..$HostCount | ForEach-Object {
+            # Create the custom name using the values of the other parameters
             $CustomName = $Prefix + "-" + $DevType + $OSTxt + $DFunction + $('{0:d3}' -f [int]$_)
+            # Add the custom name to the output array
             $Output += $CustomName
         }
         # Create Device Name
     }
     end {
+        # Return the output array
         return $Output
     }
 }