+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added comments and help blocks to various functions
+- Added `Submit-FTPUpload` as public function
+
+## [0.1.6] - 2023-02-20
+
+### Added
+
 - Added comment help block to `Get-ADUserLogonAudit`
 ## [0.1.5] - 2023-02-20